Move one more bg job from BL project
